requirements.txt
groq
streamlit
python-dotenv
langchain
langchain-community
faiss-cpu
pypdf
sentence-transformers
